requirements.txt
aiofiles==23.2.1
aiogram==3.5.0
aiohttp==3.9.0
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.11.2
bleach==4.1.0
cachetools==5.5.1
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==43.0.1
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
deepdiff==8.0.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
Django==5.0.12
django-botmanager==0.2.21
django-countries==7.6.1
django-debug-toolbar==5.0.1
django-dirtyfields==1.9.3
django-filter==23.5
django-modelcluster==6.3
django-permissionedforms==0.1
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs-exporter==2.1.7
effdet==0.4.1
emoji==2.12.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fabric==3.2.2
fake-useragent==1.5.1
filelock==3.17.0
filetype==1.2.0
flake8==7.1.2
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.4.1
fsspec==2025.2.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-vision==3.10.0
googleapis-common-protos==1.67.0
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.4
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.10.1
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.28.1
humanfriendly==10.0
hyperframe==5.2.0
idna==3.4
invoke==2.2.0
iopath==0.1.10
Jinja2==3.1.5
jiter==0.5.0
joblib==1.4.2
jsonfield==3.1.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kiwisolver==1.4.8
l18n==2021.3
laces==0.1.1
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langdetect==1.0.9
langsmith==0.1.114
layoutparser==0.3.4
lxml==5.3.0
magic-filter==1.0.12
Markdown==3.7
MarkupSafe==3.0.2
marshmallow==3.22.0
matplotlib==3.10.0
mccabe==0.7.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.1
openai==1.43.0
opencv-python==4.11.0.86
openpyxl==3.1.5
orderly-set==5.2.2
orjson==3.10.7
packaging==24.1
pandas==2.2.3
paramiko==3.4.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pi_heif==0.21.0
pikepdf==9.5.2
pillow==10.4.0
pillow_heif==0.18.0
pipdeptree==2.25.0
portalocker==3.1.1
promise==2.3
proto-plus==1.26.0
protobuf==5.29.3
psutil==5.8.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycodestyle==2.12.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pyflakes==3.2.0
PyNaCl==1.5.0
pyparsing==3.2.1
pypdf==4.3.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.1
pytz==2024.1
PyYAML==6.0.2
rapidfuzz==3.9.7
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
rsa==4.9
safetensors==0.5.2
scipy==1.15.2
setproctitle==1.1.10
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
sqlparse==0.5.1
sympy==1.13.1
tabulate==0.9.0
telepath==0.3.1
tenacity==8.5.0
text-unidecode==1.3
tiktoken==0.7.0
timm==1.0.14
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.66.5
transformers==4.48.3
triton==3.2.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.15.9
unstructured-client==0.25.7
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.13
urllib3==2.2.2
wagtail==6.2.1
wagtail-markdown==0.12.1
webencodings==0.5.1
Willow==1.9.0
wrapt==1.16.0
yarl==1.9.11
